vportscan_detector: added scan_aggregator (module in python) which aggregates alerts from vportscan_detector in given time window
- note: changed PORT_CNT field in vportscan_detector output template from uint8 to uint32

+++ vportscan_detector/vportscan_detector.c
@@ -87,7 +87,7 @@ UR_FIELDS (
    uint8 EVENT_TYPE,
    time TIME_FIRST,
    time TIME_LAST,
-   uint8 PORT_CNT,
+   uint32 PORT_CNT
 )
 
 trap_module_info_t *module_info = NULL;
